Dealing with errors
Enter an integer value: 0,3
There was an error. Here is more info...
java.util.InputMismatchException
Enter a short value (must be between -32 768 and 32 767): game
The value is the wrong type. May be it is too big or small?
Thank you for your answer (finally block)!
Thank you for your answer (after try and catch blocks)!